Split monolithic files — recipe/ package root (3 files)
Split three oversized files into cohesive sub-modules:

- _cmd_rpc.py (865 lines) → _cmd_rpc_guards.py, _cmd_rpc_merge.py, _cmd_rpc_issues.py
- contracts.py (764 lines) → _contracts_types.py, _contracts_manifest.py, _contracts_card.py, _contracts_staleness.py
- _api.py (731 lines) → _api_cache.py, _api_listing.py

Each original file is preserved as a thin re-export facade to maintain
all existing import paths. Globals for cache/staleness moved to
_api_cache.py with dotted module access from load_and_validate.

Test monkeypatch targets updated for 5 test files.
