+++ tests/infra/test_schema_version_convention.py
@@ -160,8 +160,8 @@ _LEGACY_JSON_WRITES: set[tuple[str, int]] = {
     ("src/autoskillit/planner/consolidation.py", 333),
     # planner/manifests.py — finalize_wp_manifest: wp_index.json rebuild (list payload)
     ("src/autoskillit/planner/manifests.py", 269),
-    # _cmd_rpc.py — emit_fallback_map: BEM fallback execution map (recipe-internal)
-    ("src/autoskillit/recipe/_cmd_rpc.py", 588),
+    # _cmd_rpc_issues.py — emit_fallback_map: BEM fallback execution map (recipe-internal)
+    ("src/autoskillit/recipe/_cmd_rpc_issues.py", 83),
 }
