+++ tests/recipe/test_recipe_temp_substitution.py
@@ -135,8 +135,9 @@ def test_load_and_validate_different_temp_dir_relpath_produces_different_content
     monkeypatch,
 ) -> None:
     import autoskillit.recipe._api as api_mod
+    import autoskillit.recipe._api_cache as cache_mod
 
-    monkeypatch.setattr(api_mod, "_LOAD_CACHE", {})
+    monkeypatch.setattr(cache_mod, "_LOAD_CACHE", {})
     _setup_project_recipe_with_placeholder(tmp_path, "temp-placeholder-test")
 
     result1 = api_mod.load_and_validate(
